requirements.txt
wheel==0.35.1
google-api-python-client==1.12.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
PyJWT==1.7.1
slackclient==2.9.1
gspread==3.6.0
clint==0.5.1
